Created menu template for adding and deleting books

+++ src/main/java/com/infoshareacademy/BookService.java
@@ -30,7 +30,7 @@ public class BookService {
     setFavouriteForBooks();
   }
 
-  public void deleteBook(int id) {
+  public void deleteBook(Long id) {
     BookRepository.getBooks().removeIf(b -> b.getId() == id);
   }
